geoscience/dark-side/Wu Shaodong/requirements.txt
albucore==0.0.21
albumentations==1.4.23
efficientnet-pytorch==0.7.1
huggingface-hub==0.27.0
mmcv==2.1.0
mmengine==0.10.5
mmsegmentation==1.2.2
natsort==8.4.0
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pandas==2.2.3
pillow==10.3.0
segmentation_models_pytorch==0.3.4
timm==0.9.7
scikit-learn
natsort
ftfy
regex
torch==2.1.2+cu118
torchvision==0.16.2+cu118
